requirements.txt
opencv-python~=4.11.0.86
pytest~=8.3.5
numpy~=2.2.6
requests~=2.32.3
pyserial~=3.5